Toch nog een update van het ERD...

+++ src/main/resources/Sprint2/Database Scripts/20211216_Create_Schema_Tables_Script_TheVault_Definitief.sql
@@ -42,7 +42,7 @@ SHOW WARNINGS;
 CREATE TABLE IF NOT EXISTS `The_Vault`.`rekening` (
   `rekeningId` INT NOT NULL AUTO_INCREMENT,
   `iban` VARCHAR(18) NOT NULL,
-  `saldo` BIGINT(255) NOT NULL,
+  `saldo` DECIMAL(50,10) NOT NULL,
   PRIMARY KEY (`rekeningId`))
 ENGINE = InnoDB;
 
@@ -117,7 +117,7 @@ SHOW WARNINGS;
 CREATE TABLE IF NOT EXISTS `The_Vault`.`asset` (
   `gebruikerId` INT NOT NULL,
   `cryptomuntId` INT NOT NULL,
-  `aantal` DECIMAL(50) NOT NULL,
+  `aantal` DECIMAL(50,10) NOT NULL,
   PRIMARY KEY (`cryptomuntId`, `gebruikerId`),
   CONSTRAINT `heeftInPortefeuille`
     FOREIGN KEY (`gebruikerId`)
@@ -144,11 +144,11 @@ DROP TABLE IF EXISTS `The_Vault`.`transactie` ;
 SHOW WARNINGS;
 CREATE TABLE IF NOT EXISTS `The_Vault`.`transactie` (
   `transactieId` INT NOT NULL AUTO_INCREMENT,
-  `aantal` DECIMAL(50) NOT NULL,
+  `aantal` DECIMAL(50,10) NOT NULL,
   `momentTransactie` DATETIME NOT NULL,
   `koperGebruikerId` INT NOT NULL,
   `cryptomuntId` INT NOT NULL,
-  `bedrag` DOUBLE NOT NULL,
+  `bedrag` DECIMAL(50,10) NOT NULL,
   `verkoperGebruikerId` INT NOT NULL,
   PRIMARY KEY (`transactieId`),
   CONSTRAINT `kooptMunt`
@@ -208,7 +208,7 @@ DROP TABLE IF EXISTS `The_Vault`.`dagkoersCrypto` ;
 
 SHOW WARNINGS;
 CREATE TABLE IF NOT EXISTS `The_Vault`.`dagkoersCrypto` (
-  `waardeCrypto` DOUBLE NOT NULL,
+  `waardeCrypto` DECIMAL(50,10) NOT NULL,
   `datum` DATE NOT NULL,
   `cryptomuntId` INT NOT NULL,
   `cryptowaardeId` VARCHAR(15) NOT NULL,
